requirements.txt
jupyterlab
tensorflow-gpu>=2.0.0-rc0
numpy